# backend/openapi.postman.yaml
openapi: 3.0.3
info:
  title: dAstIll Backend API
  version: 0.1.0
  description: |
    Import this file into Postman as an OpenAPI definition.

    Notes:
    - `GET /api/health` is public.
    - All other `/api/*` endpoints require `x-dastill-proxy-auth`.
    - Some destructive routes additionally require `x-dastill-role: operator`.
    - SSE endpoints return `text/event-stream`.
servers:
  - url: http://localhost:3001
    description: Local backend
...
security:
  - ProxyAuth: []
paths:
...
  /api/workspace/bootstrap:
    get:
      tags: [Workspace]
      summary: Bootstrap workspace data for the UI
      parameters:
        - $ref: "#/components/parameters/SelectedChannelId"
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
        - $ref: "#/components/parameters/IncludeShorts"
        - $ref: "#/components/parameters/VideoType"
        - $ref: "#/components/parameters/Acknowledged"
        - $ref: "#/components/parameters/QueueOnly"
        - $ref: "#/components/parameters/QueueTab"
      responses:
        "200":
          description: Workspace bootstrap payload
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkspaceBootstrapPayload"
        default:
          $ref: "#/components/responses/StringError"
  /api/channels:
    get:
      tags: [Channels]
      summary: List channels
      responses:
        "200":
          description: Channels
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Channel"
        default:
          $ref: "#/components/responses/StringError"
    post:
      tags: [Channels]
      summary: Add a channel
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddChannelRequest"
      responses:
        "201":
          description: Created channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        default:
          $ref: "#/components/responses/StringError"
  /api/channels/{id}:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    get:
      tags: [Channels]
      summary: Get channel
      responses:
        "200":
          description: Channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        default:
          $ref: "#/components/responses/StringError"
    put:
      tags: [Channels]
      summary: Update channel sync settings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateChannelRequest"
      responses:
        "200":
          description: Updated channel
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Channel"
        default:
          $ref: "#/components/responses/StringError"
    delete:
      tags: [Channels]
      summary: Delete channel
      description: "Requires `x-dastill-role: operator`."
      security:
        - ProxyAuth: []
          OperatorRole: []
      responses:
        "204":
          description: Channel deleted
        default:
          $ref: "#/components/responses/StringError"
  /api/channels/{id}/sync-depth:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    get:
      tags: [Channels]
      summary: Get channel sync depth
      responses:
        "200":
          description: Sync depth
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SyncDepthPayload"
        default:
          $ref: "#/components/responses/StringError"
  /api/channels/{id}/snapshot:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/IncludeShorts"
      - $ref: "#/components/parameters/VideoType"
      - $ref: "#/components/parameters/Acknowledged"
      - $ref: "#/components/parameters/QueueOnly"
      - $ref: "#/components/parameters/QueueTab"
    get:
      tags: [Channels]
      summary: Get channel snapshot with filtered videos
      responses:
        "200":
          description: Channel snapshot
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelSnapshotPayload"
        default:
          $ref: "#/components/responses/StringError"
  /api/channels/{id}/refresh:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
    post:
      tags: [Channels]
      summary: Refresh latest videos for a channel
      responses:
        "200":
          description: Refresh result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideosAddedResponse"
        default:
          $ref: "#/components/responses/StringError"
  /api/channels/{id}/backfill:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
      - name: limit
        in: query
        schema:
          type: integer
          minimum: 1
          maximum: 100
          default: 15
      - name: until
        in: query
        description: RFC3339 timestamp
        schema:
          type: string
          format: date-time
    post:
      tags: [Channels]
      summary: Backfill older videos for a channel
      responses:
        "200":
          description: Backfill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ChannelBackfillResponse"
        default:
          $ref: "#/components/responses/StringError"
  /api/channels/{id}/videos:
    parameters:
      - $ref: "#/components/parameters/ChannelId"
      - $ref: "#/components/parameters/Limit"
      - $ref: "#/components/parameters/Offset"
      - $ref: "#/components/parameters/IncludeShorts"
      - $ref: "#/components/parameters/VideoType"
      - $ref: "#/components/parameters/Acknowledged"
      - $ref: "#/components/parameters/QueueOnly"
      - $ref: "#/components/parameters/QueueTab"
    get:
      tags: [Videos]
      summary: List videos for a channel
      responses:
        "200":
          description: Videos
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Video"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [Videos]
      summary: Get video
      responses:
        "200":
          description: Video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/info:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [Videos]
      summary: Get stored video info
      responses:
        "200":
          description: Video info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoInfo"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/info/ensure:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Videos]
      summary: Fetch video info if missing or stale
      responses:
        "200":
          description: Video info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoInfo"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/info/backfill:
    post:
      tags: [Videos]
      summary: Backfill video info rows
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 1000
            default: 100
        - name: force
          in: query
          schema:
            type: boolean
            default: false
        - name: heal_placeholders
          in: query
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Backfill result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VideoInfoBackfillResponse"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/acknowledged:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    put:
      tags: [Videos]
      summary: Update acknowledged state
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAcknowledgedRequest"
      responses:
        "200":
          description: Updated video
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Video"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/transcript:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [Content]
      summary: Get transcript
      responses:
        "200":
          description: Transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transcript"
        default:
          $ref: "#/components/responses/StringError"
    put:
      tags: [Content]
      summary: Save transcript manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentRequest"
      responses:
        "200":
          description: Transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transcript"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/transcript/ensure:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Content]
      summary: Generate or fetch transcript
      responses:
        "200":
          description: Transcript
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transcript"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/transcript/clean:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Content]
      summary: Clean transcript formatting with AI
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentRequest"
      responses:
        "200":
          description: Cleaning result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CleanTranscriptResponse"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/summary:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    get:
      tags: [Content]
      summary: Get summary
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        default:
          $ref: "#/components/responses/StringError"
    put:
      tags: [Content]
      summary: Save summary manually
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateContentRequest"
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/summary/ensure:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Content]
      summary: Generate or fetch summary
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/summary/regenerate:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Content]
      summary: Force summary regeneration
      responses:
        "200":
          description: Summary
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Summary"
        default:
          $ref: "#/components/responses/StringError"
  /api/videos/{id}/reset:
    parameters:
      - $ref: "#/components/parameters/VideoId"
    post:
      tags: [Content]
      summary: Reset transcript and summary processing state
      responses:
        "204":
          description: Video reset
        default:
          $ref: "#/components/responses/StringError"
  /api/search:
    get:
      tags: [Search]
      summary: Search transcripts and summaries
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
        - name: source
          in: query
          schema:
            $ref: "#/components/schemas/SearchSourceFilter"
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 25
            default: 8
        - name: channel_id
          in: query
          schema:
            type: string
        - name: mode
          in: query
          schema:
            $ref: "#/components/schemas/SearchExecutionMode"
      responses:
        "200":
          description: Search results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchResponsePayload"
        default:
          $ref: "#/components/responses/StringError"
  /api/search/status:
    get:
      tags: [Search]
      summary: Current search indexing status
      responses:
        "200":
          description: Search status
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchStatusPayload"
        default:
          $ref: "#/components/responses/StringError"
  /api/search/status/stream:
    get:
      tags: [Search]
      summary: Stream search indexing status updates
      responses:
        "200":
          description: Server-sent events of `SearchStatusPayload`
          content:
            text/event-stream:
              schema:
                type: string
              examples:
                sample:
                  value: |
                    data: {"available":true,"model":"nomic-embed-text","dimensions":512,"pending":0,"indexing":0,"ready":12,"failed":0,"total_sources":12,"total_chunk_count":44,"embedded_chunk_count":44,"vector_index_ready":true,"retrieval_mode":"hybrid_ann"}

        default:
          $ref: "#/components/responses/StringError"
  /api/search/rebuild:
    post:
      tags: [Search]
      summary: Rebuild the search projection
      responses:
        "202":
          description: Rebuild accepted
        default:
          $ref: "#/components/responses/StringError"
...
  /api/chat/conversations/{id}/stream:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
    get:
      tags: [Chat]
      summary: Reconnect to an active chat SSE stream
      responses:
        "200":
          description: Chat response stream
          content:
            text/event-stream:
              schema:
                type: string
        default:
          $ref: "#/components/responses/StringError"
  /api/chat/conversations/{id}/cancel:
    parameters:
      - $ref: "#/components/parameters/ConversationId"
    post:
      tags: [Chat]
      summary: Cancel an active chat response
      responses:
        "202":
          description: Cancellation accepted
        default:
          $ref: "#/components/responses/StringError"
...
components:
  securitySchemes:
    ProxyAuth:
      type: apiKey
      in: header
      name: x-dastill-proxy-auth
      description: Required for all protected endpoints.
    OperatorRole:
      type: apiKey
      in: header
      name: x-dastill-role
      description: Set to `operator` for operator-only endpoints.
    ClientIp:
      type: apiKey
      in: header
      name: x-dastill-client-ip
      description: Optional stable client identifier used for rate limiting and anonymous chat quota.
  parameters:
...
    SelectedChannelId:
      name: selected_channel_id
      in: query
      schema:
        type: string
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 0
        maximum: 100
        default: 20
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
    IncludeShorts:
      name: include_shorts
      in: query
      schema:
        type: boolean
        default: true
    VideoType:
      name: video_type
      in: query
      schema:
        $ref: "#/components/schemas/VideoTypeFilter"
    Acknowledged:
      name: acknowledged
      in: query
      schema:
        type: boolean
    QueueOnly:
      name: queue_only
      in: query
      schema:
        type: boolean
...